logging: Add custom logging-level VERBOSE
This commit fixes verbose() using the DEBUG logging-level, and instead
uses a custom logging level VERBOSE (INFO-1). This logging level is
defined in param.parameterized. To provide all logging-levels in one
module namespace, the logging-levels defined in the logging module are
imported into the param.parameterized namespace.

The following example shows how to initialize the root-logger with the
VERBOSE log-level and a custom log-format:
  import logging
  import param.parameterized
  logging.basicConfig(level=param.parameterized.VERBOSE, format='[%(levelname)s] %(message)s')