Provide Version fix when project is copied to foreign GIT repository
Implement fix for Version to not cause Exception when the project's
source code is copied to a foreign GIT repository.

The fix is to specify an optional project-name, passed to Version. If a
name is provided, the code obtaining version information from GIT first
tries to verify we are in the correct GIT repository. A simple heuristic
was implemented, which checks the output of 'git remote -v' for the name
of the project.