Added options to automatically document parameters in class docstrings
Introduced option to automatically document parameter names and types on
the first line of the Parameterized class docstrings. This allows TAB
completion of parameter keywords in the IPython environment.

In addition, the output of ParamPager may be optionally appended to the
class docstring, supplying a complete description of all available
parameters. Implements PR #72