Restored previous (pre-1.3.0) version of script_repr.
The new pprint is still available, but is marked experimental. Added pprint_reg (like script_repr reg) and filled with any entries necessary to override those in script_repr_reg. This has all caused the code to balloon, but we need to support the original script_repr while also adding the improved (but-not-quite-complete-replacement-for-script_repr) pprint for holoviews.