Defined namespace and implementation on Parameters object (#230)
Moves most of the code of `Parameterized` onto a `param` object of type `Parameters`, which also acts as a namespace. The goal is to make this the face of the public API, cleaning up the namespace of user classes.

The new API is likely to change as we clean it up, but the previous API remains available, so existing code should not be affected. The previous API will be deprecated at some point in the future; `Parameters._disable_stubs` can be used to control deprecation warnings/errors/silence.

Docstrings of methods on Parameterized that have moved to Parameters point readers to the new methods. 

Existing tests have been copied into an `API0` subdirectory, while tests in `API1` use the Parameters namespace. (This wholesale copying could probably be improved in the future.)

Ideally, `dir()` would already return only the cleaned up namespace (i.e. not the deprecated methods on `Parameterized`), but this has not been done yet and may not be easy.